print keypress nativeScancode, nativeVirtualKey, nativeModifiers; and add use 4 byte storage keypress code

+++ project/RemoteControl/CClient/src/CControlWnd.h
@@ -83,7 +83,7 @@ protected:
     void wheelEvent(QWheelEvent *);
     void keyPressEvent(QKeyEvent *);
     void keyReleaseEvent(QKeyEvent *);
-    uchar translateKey(int key);
+    int translateKey(int key);
 };
 
 #endif // CONTROLPANEL_H
